rename folder schemas
